install sleep

+++ test_actions/setup.bash
@@ -42,7 +42,9 @@ sed -i "s#SLACK_CLI_TOKEN_SECRET#${SLACK_CLI_TOKEN_SECRET}#" ${DEPLOYMENT_PATH}/
 sed -i "s#GITHUB_APP_ID_PLACEHOLDER#${GH_APP_ID}#" ${DEPLOYMENT_PATH}/overlays/test-local/config-source/github.env
 sed -i "s#GITHUB_APP_LOGIN_PLACEHOLDER#${GITHUB_APP_LOGIN_PLACEHOLDER}#" ${DEPLOYMENT_PATH}/overlays/test-local/config-source/github.env
 sed -i "s#GITHUB_INSTALLATION_ID_PLACEHOLDER#${GH_INSTALLATION_ID}#" ${DEPLOYMENT_PATH}/overlays/test-local/config-source/github.env
+if [ -e ${GH_PRIVATE_KEY_PATH} ]; then
 cp "${GH_PRIVATE_KEY_PATH}" ${DEPLOYMENT_PATH}/overlays/test-local/config-source/github_private_key.pem || true
+fi
 if [ "${GH_PRIVATE_KEY}" != "" ]; then
   cat "${GH_PRIVATE_KEY_PATH}" > ${DEPLOYMENT_PATH}/overlays/test-local/config-source/github_private_key.pem
 fi
